requirements.txt
accelerate==0.21.0
anthropic==0.25.6
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
boto3==1.34.89
botocore==1.34.89
bs4==0.0.2
datasets==2.19.0
deepspeed==0.12.6
filelock==3.13.4
flash-attn==2.5.8
huggingface-hub==0.22.2
matplotlib==3.8.4
numpy==1.26.4
openai==1.23.2
pandas==2.2.2
peft==0.10.0
pillow==10.3.0
protobuf==4.25.3
pyarrow==16.0.0
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.2.2
scipy==1.13.0
tokenizers==0.15.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.2
transformers==4.37.2
triton==2.1.0
urllib3==2.2.1
uvicorn==0.29.0
vertexai==1.48.0
wandb==0.17.0
wavedrom==2.0.3.post3
